import votes from Discourse using its REST API
Instead of guessing which assembly was a poll for (using its date), we will store a list of Discourse topics containing each vote of an assembly in the config YAML (and in the database). All changes are the same for assemblies and voting sessions.

Instead of storing `poll_id` (which we don’t have access to via REST API), we will be storing a voting topic’s ID for each `Vote`. If a person voted in all polls in a topic, there will be only one entry for all these votes (we don’t really to store even a `topic_id` for activity calculations, but this simplifies debugging).

With this change, we will no longer need access to the Discourse database.